+++ frontend2/src/routes/dashboard.tsx
@@ -2,7 +2,25 @@ export default function Dashboard() {
 
   return (
     <>
-      <p>Dashboard</p>
+      <p class='text-3xl bg-green-500 p-10'>Dashboard</p>
+      <button class="text-blue-400 hover:text-blue-500 font-mono font-light py-2 px-4 border-2 border-rounded border-blue-200"
+      >
+        Button
+      </button>
+
+      <div class="i-lucide-beef text-lg" />
+      <div class="i-ph-anchor-simple-thin text-5xl" />
+
+      <div class="i-mdi-alarm text-orange-400" />
+
+      <div class="i-logos-vue text-3xl" />
+      <div class='i-logos-react text-base'></div>
+      <button class="i-carbon-sun dark:i-carbon-moon" />
+
+      <div class="i-twemoji-grinning-face-with-smiling-eyes hover:i-twemoji-face-with-tears-of-joy" />
+      <span class="i-carbon-logo-github text-lg"></span>
+      <span class="i-mdi-light-home"></span>
     </>
+
   )
 }