+++ frontend2/src/routes/home.tsx
@@ -1,7 +1,7 @@
 import { createSignal, For, Show } from 'solid-js';
-import solidLogo from '~/assets/solid.svg';
-import viteLogo from '/vite.svg';
-import '~/App.css';
+// import solidLogo from '~/assets/solid.svg';
+// import viteLogo from '/vite.svg';
+// import '~/App.css';
 
 import { lazy } from "solid-js";
 
@@ -37,14 +37,6 @@ function App() {
 
   return (
     <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} class="logo" alt="Vite logo" />
-        </a>
-        <a href="https://solidjs.com" target="_blank">
-          <img src={solidLogo} class="logo solid" alt="Solid logo" />
-        </a>
-      </div>
       <h1>Vite + Solid + DOHAC Mock API</h1>
       <div class="card">
         <button onClick={() => setCount((count) => count + 1)}>
@@ -61,7 +53,7 @@ function App() {
       </div>
 
       {/* --- Display API Data --- */}
-      <div class="api-data">
+      <div>
         <h2>Provider Data</h2>
         <Show when={loading()}>
           <p>Loading...</p>